Reverted Typescript integration
