+++ config/db.js
@@ -10,7 +10,7 @@ const connectDB = async () => {
     useUnifiedTopology: true,
     useNewUrlParser: true,
     useCreateIndex: true,
-    useFindAndModify: false
+    useFindAndModify: false,
   });
   console.log(`MongoDB Connected: ${conn.connection.host}`);
 };
