--- src/config/config.env
@@ -1,3 +1,0 @@
-MONGO_URL="mongodb+srv://tharising:tharising@esportshaven-vpww7.mongodb.net/data?retryWrites=true&w=majority"
-NODE_ENV="PROD"
-JWT_SECRET="SecretKey!!"
