--- tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "incremental": true,
-    "target": "es2015",
-    "module": "esnext",
-    "lib": [
-      "es2018",
-      "dom"
-    ],
-    "allowJs": true,
-    "declaration": false,
-    "sourceMap": true,
-    "outDir": "./src",
-    "importHelpers": true,
-    "downlevelIteration": true,
-    "moduleResolution": "node",
-    "baseUrl": "./",
-    "typeRoots": [
-      "node_modukes/@types"
-    ],
-    "esModuleInterop": true,
-    "experimentalDecorators": true,
-    "forceConsistentCasingInFileNames": true
-  }
-}
